Moves logging module to autoconfigure
Still some moving required.
