+++ spring-cloud-gcp-autoconfigure/src/main/java/org/springframework/cloud/gcp/autoconfigure/logging/TraceIdExtractor.java
@@ -14,29 +14,22 @@
  *  limitations under the License.
  */
 
-package org.springframework.cloud.gcp.logging;
+package org.springframework.cloud.gcp.autoconfigure.logging;
 
 import javax.servlet.http.HttpServletRequest;
 
 /**
- * Extracts trace IDs from HTTP requests using the x-cloud-trace-context header.
+ * An extractor that can provide a trace ID from an HTTP request.
  *
  * @author Chengyuan Zhao
  */
-public class XCloudTraceIdExtractor implements TraceIdExtractor {
+public interface TraceIdExtractor {
 
-	public static final String X_CLOUD_TRACE_HEADER = "x-cloud-trace-context";
-
-	@Override
-	public String extractTraceIdFromRequest(HttpServletRequest req) {
-		String traceId = req.getHeader(X_CLOUD_TRACE_HEADER);
-
-		if (traceId != null) {
-			int slash = traceId.indexOf('/');
-			if (slash >= 0) {
-				traceId = traceId.substring(0, slash);
-			}
-		}
-		return traceId;
-	}
+	/**
+	 * Extract trace ID from the HTTP request.
+	 *
+	 * @param req The HTTP servlet request.
+	 * @return The trace ID or null, if none found.
+	 */
+	String extractTraceIdFromRequest(HttpServletRequest req);
 }
