+++ spring-cloud-gcp-autoconfigure/src/main/resources/logging/logback-appender.xml
