+++ spring-cloud-gcp-autoconfigure/src/main/resources/logging/logback-json-appender.xml
@@ -7,7 +7,7 @@ Stackdriver json-log format provided for import.
 
     <appender name="CONSOLE_JSON" class="ch.qos.logback.core.ConsoleAppender">
         <encoder class="ch.qos.logback.core.encoder.LayoutWrappingEncoder">
-            <layout class="org.springframework.cloud.gcp.logging.StackdriverJsonLayout">
+            <layout class="org.springframework.cloud.gcp.autoconfigure.logging.StackdriverJsonLayout">
                 <projectId>${projectId}</projectId>
 
                 <!--<includeTraceId>true</includeTraceId>-->
